레포지터리 인터페이스
